CRUD methods for the identified objects of the project

+++ controllers/productController.js
@@ -1,4 +1,5 @@
 const productRepository = require("../repositories/productRepository");
+const categoryRepository = require("../repositories/categoryRepository");
 const asyncHandler = require("express-async-handler");
 const { validationResult } = require("express-validator");
 
@@ -76,13 +77,94 @@ exports.product_update = asyncHandler(async (req, res, next) => {
 });
 
 exports.product_delete = asyncHandler(async (req, res, next) => {
+  var errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(404).json({ errors: errors.array() });
+  }
+
   const { productId } = req.params;
 
-  //check if product exists
+  const product = await productRepository.productById(productId);
 
-  await productRepository.deleteProduct(productId);
+  if (!product) {
+    return res.status(404).json({
+      error: true,
+      msg: "producto no fue encontrado en la base de datos",
+    });
+  }
 
-  //retrive product again to check if enable is false
+  const result = await productRepository.deleteProduct(productId);
 
-  return res.status(200).json({ msg: "product disabled" });
+  res.status(200).json(result);
+});
+
+exports.createProductByCategory = asyncHandler(async (req, res, next) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
+  const { productId } = req.params;
+  const { categoryId } = req.params;
+
+  const product = await productRepository.productById(productId);
+
+  if (!product) {
+    return res.status(404).json({
+      error: true,
+      msg: `El producto con id ${productId} no fue encontrado`,
+    });
+  }
+
+  const category = await categoryRepository.categoryById(categoryId);
+
+  if (!category) {
+    return res.status(404).json({
+      error: true,
+      msg: `La categoria con el id ${categoryId} no fue encontrada`,
+    });
+  }
+
+  const productByCategory = await productRepository.createProductByCategory(
+    product,
+    category
+  );
+
+  res.status(201).json(productByCategory);
+});
+
+exports.deleteProductByCategory = asyncHandler(async (req, res, next) => {
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
+
+  const { productId, categoryId } = req.params;
+
+  const product = await productRepository.productById(productId);
+  const category = await categoryRepository.categoryById(categoryId);
+
+  if (!product) {
+    return res.status(400).json({
+      errors: true,
+      msg: "El producto no se existe en la base de datos",
+    });
+  }
+
+  if (!category) {
+    return res.status(400).json({
+      errors: true,
+      msg: "La categoria no existe en la base de datos",
+    });
+  }
+
+  const result = await productRepository.removeProductByCategory(
+    product,
+    category
+  );
+
+  res.status(200).json(result);
 });
